Action for creating a ResetKey for members by email.
A POST to 'forgot-password' with an email, will look up the
member by email, and if the member exists, calls forgot_password!
on that member. This method creates the ResetKey, and will email
the user with instructions on using it. The page rendered is a
simple "Email Sent" page.

If the email is not that of a member, then it will redirect back
and flash an error message.