Let's use Bootstrap
I added in Bootstrap 3 (without a gem, since I've tried two gems are
they're both not letting us have nice things like navbars).

Bootstrap gives us:

* Normalize (which we were including anyway)
* A bunch of sane defaults for things like form defaults, pagination,
  and other standard UI elements
* A (responsive) grid system
* Cross-browser compatability

Like, legit, let's leverage Bootstrap. It'd be stupid not to. Let's
not invent a grid system or form elements, let's just customize the
crap out of Boostrap -- once we've got something that *works*.

Changes you might not like:

The navbar is now the (infamous) Bootstrap navbar. That should (unless
we change our minds) get styled to look like we wanted in our mockups,
but this'll get us going for now. Let's style on top of Bootstrap
rather than trying to reinvent the wheel.

The bigpic on every page is no longer a random image. I wasn't sure
how to get this working since I'm setting the image using CSS.
What we could do is add an ID to the body based on the page and use a
set image for each page, and change the text in the #pitch div based
on the page.